+++ Arma3ObjectBuilder/io/__init__.py
@@ -1,9 +1,10 @@
 from . import binary_handler
 from . import data_p3d
+from . import data_rap
 from . import data_rtm
 from . import export_asc
 from . import export_p3d
 from . import export_rtm
 from . import import_asc
+from . import import_mcfg
 from . import import_p3d
-from . import import_rap
