+++ Arma3ObjectBuilder/io/data_rap.py
